Refactor current mock store implementation
Moving this to another file so that multiple tests can use it. In that
dedicated file will go all mocks for any tests we're gonna run.
